+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-08-06
+
+### Added
+
+- Add function get_delay_timers.
+- Extend wtw.yaml with delay timer MQTT topics.
+
+### Fixed
+
+- Function get_fan_status had the wrong command in the documentation part.
+- Reformatted the code using Black.
+
 ## [1.0.1] - 2023-08-05
 
 ### Added